requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-autoslug==1.9.8
django-cleanup==6.0.0
django-uuslug==2.0.0
Pillow==9.2.0
python-slugify==6.1.2
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
